Move burrow evm tp to processor/ dir
This more closely follows the structure of the intkey folder and helps show a
clearer separation between the processor and client packages.
