+++ bin/run_go_fmt
@@ -10,8 +10,8 @@ $top_dir/sdk/go/src/sawtooth_sdk/processor
 $top_dir/sdk/go/src/sawtooth_sdk/client
 $top_dir/sdk/examples/intkey_go/src/sawtooth_intkey
 $top_dir/sdk/examples/intkey_go/src/sawtooth_intkey/handler
-$top_dir/families/burrow_evm/src/sawtooth_burrow_evm
-$top_dir/families/burrow_evm/src/sawtooth_burrow_evm/handler
+$top_dir/families/burrow_evm/src/sawtooth_burrow_evm/processor
+$top_dir/families/burrow_evm/src/sawtooth_burrow_evm/processor/handler
 $top_dir/families/burrow_evm/src/sawtooth_burrow_evm/client
 "
